Count tweets by citycode
Import geonamescache to get official names of location and get the keys as citycode
count the tweets based on the citycode colected
Create 2 dictionaries: one is citycode and tweetcount, another is citycode and city name, countrycode and states. Create 2 csv files
Merge two of them using pandas and add headers.